requirements.txt
# Core dependencies
yt-dlp>=2023.7.6
openai>=0.27.0
python-dotenv>=1.0.0

# Google APIs for YouTube
google-api-python-client>=2.95.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0

# Web automation for Rumble
selenium>=4.10.0
webdriver-manager>=3.8.6

# Video processing
Pillow>=10.0.0

# Data handling
requests>=2.31.0

# Optional: Additional tools
psutil>=5.9.0  # For system monitoring
tqdm>=4.65.0   # For progress bars